Key signature fix
Signatures generated from files work only with other signatures generated from files!
Because of an error with \n handling
This means that we should handle HS256 as generated from server and RS256 as generated from file.
When we apply the RS256 to HS256 vuln we should act like it's generated from file