Fix intendation/scope for block responsible of jku-var exchange.
I belive this addresses the problem described in issue #90.
To reproduce the undefined var in line it I had to run in the scan mode
"-M at" and you need a JWT that defines "jku" field. In that case the try
block before succeds. That lead to the "jku" variable never to be defined.